Correct dictionary names
Plover has a bug where systems having the same name for dictionaries can
cause it to load the main.json, for example, of another system if a new
system tries to use that name for its default dictionary. As a result,
we should always prefix our new system default dictionaries.